add unit test, and fix some issues
