+++ .gitignore
@@ -480,6 +480,12 @@ packages/app-desktop/gui/NoteEditor/utils/clipboardUtils.test.js.map
 packages/app-desktop/gui/NoteEditor/utils/contextMenu.d.ts
 packages/app-desktop/gui/NoteEditor/utils/contextMenu.js
 packages/app-desktop/gui/NoteEditor/utils/contextMenu.js.map
+packages/app-desktop/gui/NoteEditor/utils/contextMenu.test.d.ts
+packages/app-desktop/gui/NoteEditor/utils/contextMenu.test.js
+packages/app-desktop/gui/NoteEditor/utils/contextMenu.test.js.map
+packages/app-desktop/gui/NoteEditor/utils/contextMenuUtils.d.ts
+packages/app-desktop/gui/NoteEditor/utils/contextMenuUtils.js
+packages/app-desktop/gui/NoteEditor/utils/contextMenuUtils.js.map
 packages/app-desktop/gui/NoteEditor/utils/index.d.ts
 packages/app-desktop/gui/NoteEditor/utils/index.js
 packages/app-desktop/gui/NoteEditor/utils/index.js.map
