+++ packages/app-desktop/gui/NoteEditor/utils/contextMenu.ts
@@ -1,7 +1,9 @@
 import ResourceEditWatcher from '@joplin/lib/services/ResourceEditWatcher/index';
 import { _ } from '@joplin/lib/locale';
 import { copyHtmlToClipboard } from './clipboardUtils';
+import { svgUriToPng } from './resourceHandling';
 import bridge from '../../../services/bridge';
+import { ContextMenuItemType, ContextMenuOptions, ContextMenuItems, resourceInfo, tempResourceToDataUri } from './contextMenuUtils';
 const Menu = bridge().Menu;
 const MenuItem = bridge().MenuItem;
 import Resource from '@joplin/lib/models/Resource';
@@ -13,68 +15,12 @@ const fs = require('fs-extra');
 const { clipboard } = require('electron');
 const { toSystemSlashes } = require('@joplin/lib/path-utils');
 
-export enum ContextMenuItemType {
-	None = '',
-	Image = 'image',
-	Resource = 'resource',
-	Text = 'text',
-	Link = 'link',
-}
-
-export interface ContextMenuOptions {
-	itemType: ContextMenuItemType;
-	resourceId: string;
-	tempResource: {
-		data: any; // string or Uint8Array
-		mime: string;
-		filename: string;
-	};
-	linkToCopy: string;
-	textToCopy: string;
-	htmlToCopy: string;
-	insertContent: Function;
-	isReadOnly?: boolean;
-	isTemp?: boolean;
-}
-
-interface ContextMenuItem {
-	label: string;
-	onAction: Function;
-	isActive: Function;
-}
-
-interface ContextMenuItems {
-	[key: string]: ContextMenuItem;
-}
-
-async function resourceInfo(options: ContextMenuOptions): Promise<any> {
-	const resource = options.resourceId ? await Resource.load(options.resourceId) : null;
-	const filePath = resource ? Resource.fullPath(resource) : null;
-	const filename = resource ? (resource.filename ? resource.filename : resource.title) : options.tempResource && options.tempResource.filename ? options.tempResource.filename : '';
-	const getCopyPath = () => {
-		if (options.tempResource) {
-			return tempResourceToDataUri(options.tempResource);
-		}
-		return filePath;
-	};
-	return { resource, filePath, isTemp: !!options.tempResource, filename, getCopyPath };
-}
-
 function handleCopyToClipboard(options: ContextMenuOptions) {
 	if (options.textToCopy) {
 		clipboard.writeText(options.textToCopy);
 	} else if (options.htmlToCopy) {
 		copyHtmlToClipboard(options.htmlToCopy);
 	}
-}
-
-function tempResourceToDataUri(tempResource: ContextMenuOptions['tempResource']): string {
-	const mime = tempResource.mime;
-	if (mime.startsWith('image/svg')) {
-		const uri = `data:image/svg+xml;base64,${Buffer.from(tempResource.data).toString('base64')}`;
-		return uri;
-	}
-	return null;
 }
 
 async function saveResource(options: ContextMenuOptions) {
@@ -84,34 +30,6 @@ async function saveResource(options: ContextMenuOptions) {
 	if (options.isTemp) {
 		await fs.outputFile(newFilePath, options.tempResource.data);
 	} else { await fs.copy(filePath, newFilePath); }
-}
-
-function svgToPng(svg: string, cb: Function) {
-	let canvas: HTMLCanvasElement;
-	let ctx;
-	const img = new Image();
-	img.onload = function() {
-		canvas = document.createElement('canvas');
-		canvas.width = img.width;
-		canvas.height = img.height;
-		ctx = canvas.getContext('2d');
-		ctx.drawImage(img, 0, 0, img.width, img.height, 0, 0, img.width, img.height);
-		const pngUri = canvas.toDataURL('image/png');
-		const pngBase64 = pngUri.split(',')[1];
-		const byteString = atob(pngBase64);
-		// write the bytes of the string to a typed array
-		const buff = new Uint8Array(byteString.length);
-		for (let i = 0; i < byteString.length; i++) {
-			buff[i] = byteString.charCodeAt(i);
-		}
-		cb(buff);
-	};
-	img.src = svg;
-	setTimeout(() => {
-		// Cleanup the elements
-		if (img) img.remove();
-		if (canvas) canvas.remove();
-	},3000);
 }
 
 export async function openItemById(itemId: string, dispatch: Function, hash: string = '') {
@@ -167,28 +85,27 @@ export function menuItems(dispatch: Function): ContextMenuItems {
 				await saveResource(options);
 			},
 			// We handle svg seperately as it can be saved in multiple formats
-			isActive: (itemType: ContextMenuItemType, options: ContextMenuOptions) => itemType === (ContextMenuItemType.Image || itemType === ContextMenuItemType.Resource) && (!options.isTemp || !options.tempResource.mime.startsWith('image/svg')),
+			isActive: (itemType: ContextMenuItemType, options: ContextMenuOptions) => itemType === (ContextMenuItemType.Image || itemType === ContextMenuItemType.Resource) && (!options.isTemp || !options.tempResource.mime?.startsWith('image/svg')),
 		},
 		saveAsSvg: {
 			label: _('Save as SVG'),
 			onAction: async (options: ContextMenuOptions) => {
 				await saveResource(options);
 			},
-			isActive: (itemType: ContextMenuItemType, options: ContextMenuOptions) => options.isTemp && itemType === ContextMenuItemType.Image && options.tempResource.mime.startsWith('image/svg'),
+			isActive: (itemType: ContextMenuItemType, options: ContextMenuOptions) => options.isTemp && itemType === ContextMenuItemType.Image && options.tempResource.mime?.startsWith('image/svg'),
 		},
 		saveAsPng: {
 			label: _('Save as PNG'),
 			onAction: async (options: ContextMenuOptions) => {
 				// First convert it to png then save
 				const dataUri = tempResourceToDataUri(options.tempResource);
-				svgToPng(dataUri, async (png: Uint8Array)=>{
-					options.tempResource.data = png;
-					options.tempResource.mime = 'image/png';
-					options.tempResource.filename = options.tempResource.filename.replace('.svg', '.png');
-					await saveResource(options);
-				});
+				const png = await svgUriToPng(dataUri);
+				options.tempResource.data = png;
+				options.tempResource.mime = 'image/png';
+				options.tempResource.filename = options.tempResource.filename.replace('.svg', '.png');
+				await saveResource(options);
 			},
-			isActive: (itemType: ContextMenuItemType, options: ContextMenuOptions) => options.isTemp && itemType === ContextMenuItemType.Image && options.tempResource.mime.startsWith('image/svg'),
+			isActive: (itemType: ContextMenuItemType, options: ContextMenuOptions) => options.isTemp && itemType === ContextMenuItemType.Image && options.tempResource.mime?.startsWith('image/svg'),
 		},
 		revealInFolder: {
 			label: _('Reveal file in folder'),
@@ -213,7 +130,7 @@ export function menuItems(dispatch: Function): ContextMenuItems {
 				const image = bridge().createImageFromPath(resourcePath);
 				clipboard.writeImage(image);
 			},
-			isActive: (itemType: ContextMenuItemType, options: ContextMenuOptions) => !options.isTemp && itemType === ContextMenuItemType.Image, //
+			isActive: (itemType: ContextMenuItemType, options: ContextMenuOptions) => !options.isTemp && itemType === ContextMenuItemType.Image,
 		},
 		cut: {
 			label: _('Cut'),
